added tests and moved to common util
