+++ common/util/src/main/java/org/thingsboard/common/util/geo/GeoUtil.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.thingsboard.script.api;
+package org.thingsboard.common.util.geo;
 
 import com.google.gson.JsonArray;
 import com.google.gson.JsonElement;
