+++ common/util/src/main/java/org/thingsboard/common/util/geo/RangeUnit.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.thingsboard.script.api;
+package org.thingsboard.common.util.geo;
 
 public enum RangeUnit {
     METER(1000.0), KILOMETER(1.0), FOOT(3280.84), MILE(0.62137), NAUTICAL_MILE(0.539957);
