+++ rule-engine/rule-engine-components/src/main/java/org/thingsboard/rule/engine/geo/TbGpsGeofencingFilterNodeConfiguration.java
@@ -16,9 +16,9 @@
 package org.thingsboard.rule.engine.geo;
 
 import lombok.Data;
+import org.thingsboard.common.util.geo.PerimeterType;
+import org.thingsboard.common.util.geo.RangeUnit;
 import org.thingsboard.rule.engine.api.NodeConfiguration;
-import org.thingsboard.script.api.PerimeterType;
-import org.thingsboard.script.api.RangeUnit;
 
 /**
  * Created by ashvayka on 19.01.18.
